# Conflicts:
#	src/main/kotlin/com/lykke/matching/engine/matching/MatchingEngine.kt
#	src/main/kotlin/com/lykke/matching/engine/order/process/LimitOrdersProcessor.kt
#	src/test/kotlin/com/lykke/matching/engine/services/InvalidBalanceTest.kt
#	src/test/kotlin/com/lykke/matching/engine/services/PersistenceErrorTest.kt

+++ src/main/kotlin/com/lykke/matching/engine/daos/NewOrder.kt
@@ -50,7 +50,7 @@ abstract class NewOrder(
     fun checkVolume(assetPair: AssetPair): Boolean {
         val volume = getAbsVolume()
         val minVolume = if (isStraight()) assetPair.minVolume else assetPair.minInvertedVolume
-        return minVolume == null || volume >= minVolume
+        return minVolume == null || volume >= minVolume.toDouble()
     }
 
     fun checkVolume(assetsPairsHolder: AssetsPairsHolder): Boolean {
