+++ src/main/kotlin/com/lykke/matching/engine/outgoing/messages/TradeInfo.kt
@@ -1,8 +1,9 @@
 package com.lykke.matching.engine.outgoing.messages
 
-import com.lykke.matching.engine.daos.FeeInstruction
+import com.lykke.matching.engine.daos.v2.FeeInstruction
 import com.lykke.matching.engine.daos.FeeTransfer
-import com.lykke.matching.engine.daos.fee.Fee
+import com.lykke.matching.engine.daos.fee.v2.Fee
+import java.math.BigDecimal
 import java.util.Date
 
 class TradeInfo(
@@ -13,7 +14,7 @@ class TradeInfo(
         val limitClientId: String,
         val limitVolume: String,
         val limitAsset: String,
-        val price: Double,
+        val price: BigDecimal,
         val limitOrderId: String,
         val limitOrderExternalId: String,
         val timestamp: Date,
@@ -21,6 +22,6 @@ class TradeInfo(
         val feeInstruction: FeeInstruction?,
         val feeTransfer: FeeTransfer?,
         val fees: List<Fee>,
-        val absoluteSpread: Double?,
-        val relativeSpread: Double?
+        val absoluteSpread: BigDecimal?,
+        val relativeSpread: BigDecimal?
 )
