+++ src/main/kotlin/com/lykke/matching/engine/services/BalanceUpdateService.kt
@@ -16,6 +16,7 @@ import com.lykke.utils.logging.MetricsLogger
 import org.apache.log4j.Logger
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
+import java.math.BigDecimal
 import java.util.Date
 
 @Service
@@ -41,12 +42,12 @@ class BalanceUpdateService @Autowired constructor (private val balancesHolder: B
             val reservedBalance = balancesHolder.getReservedBalance(message.clientId, message.assetId)
 
             val updated = balancesHolder.updateBalance(ProcessedMessage(messageWrapper.type, messageWrapper.timestamp!!, messageWrapper.messageId!!),
-                    message.clientId, message.assetId, message.amount)
+                    message.clientId, message.assetId, BigDecimal.valueOf(message.amount))
             messageWrapper.processedMessagePersisted = true
             if (updated) {
                 balancesHolder.sendBalanceUpdate(BalanceUpdate(message.uid.toString(),
                         MessageType.BALANCE_UPDATE.name, Date(),
-                        listOf(ClientBalanceUpdate(message.clientId, message.assetId, balance, message.amount, reservedBalance, reservedBalance)), messageWrapper.messageId!!))
+                        listOf(ClientBalanceUpdate(message.clientId, message.assetId, balance, BigDecimal.valueOf(message.amount), reservedBalance, reservedBalance)), messageWrapper.messageId!!))
             }
             messageWrapper.writeResponse(ProtocolMessages.Response.newBuilder())
 
@@ -67,7 +68,7 @@ class BalanceUpdateService @Autowired constructor (private val balancesHolder: B
             }
 
             val updated = balancesHolder.updateBalance(ProcessedMessage(messageWrapper.type, messageWrapper.timestamp!!, messageWrapper.messageId!!),
-                    message.clientId, message.assetId, message.amount)
+                    message.clientId, message.assetId, BigDecimal.valueOf(message.amount))
             messageWrapper.processedMessagePersisted = true
             if (!updated) {
                 messageWrapper.writeNewResponse(
@@ -80,7 +81,7 @@ class BalanceUpdateService @Autowired constructor (private val balancesHolder: B
             balancesHolder.sendBalanceUpdate(BalanceUpdate(message.uid,
                     MessageType.BALANCE_UPDATE.name,
                     Date(),
-                    listOf(ClientBalanceUpdate(message.clientId, message.assetId, balance, message.amount, reservedBalance, reservedBalance)),
+                    listOf(ClientBalanceUpdate(message.clientId, message.assetId, balance, BigDecimal.valueOf(message.amount), reservedBalance, reservedBalance)),
                     messageWrapper.messageId!!))
 
             messageWrapper.writeNewResponse(ProtocolMessages.NewResponse.newBuilder().setStatus(MessageStatus.OK.type))
